unspecified-css-subresources/font-face-unused.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>unused @font-face blocks load</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
@font-face {
  font-family: Ahem;
  src: url("/css/fonts/ahem/ahem.ttf?pipe=trickle(d5)");
}
</style>
<script>
var t1 = performance.timing.responseStart;

async_test(function(t) {
window.addEventListener("load", t.step_func_done(function() {
    var t2 = Date.now();
    var diff = t2 - t1;
    assert_less_than(diff, 5000, "ms after style");
  }));
}, "load doesn't wait for unused @font-face");
</script>
